Switched RequestPredicateResult to @VertxGen
