+++ vertx-web-validation/src/main/java/io/vertx/ext/web/validation/RequestPredicateResult.java
@@ -1,35 +1,33 @@
 package io.vertx.ext.web.validation;
 
-import io.vertx.codegen.annotations.DataObject;
 import io.vertx.codegen.annotations.VertxGen;
-import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.validation.impl.RequestPredicateResultImpl;
 
 /**
- * Data object representing a Request predicate result
+ * Result of a {@link RequestPredicate}
  */
 @VertxGen
-public class RequestPredicateResult {
+public interface RequestPredicateResult {
 
-  private String errorMessage;
+  /**
+   * Predicate succeeded
+   *
+   * @return
+   */
+  boolean succeeded();
 
-  private RequestPredicateResult(String exception) {
-    this.errorMessage = exception;
+  /**
+   * Get error of failure
+   *
+   * @return
+   */
+  String getErrorMessage();
+
+  static RequestPredicateResult success() {
+    return new RequestPredicateResultImpl(null);
   }
 
-  public boolean succeeded() {
-    return errorMessage == null;
+  static RequestPredicateResult failed(String message) {
+    return new RequestPredicateResultImpl(message);
   }
-
-  public String getErrorMessage() {
-    return errorMessage;
-  }
-
-  public static RequestPredicateResult success() {
-    return new RequestPredicateResult(null);
-  }
-
-  public static RequestPredicateResult failed(String message) {
-    return new RequestPredicateResult(message);
-  }
-
 }
